+++ test/context.test.ts
@@ -1,25 +1,10 @@
-import {
-  resolvesNext,
-  spy,
-} from "https://deno.land/std@0.203.0/testing/mock.ts";
-import { commands, type CommandsFlavor } from "../src/mod.ts";
-import {
-  Api,
-  assert,
-  assertRejects,
-  Chat,
-  Context,
-  describe,
-  it,
-  Message,
-  Update,
-  User,
-  UserFromGetMe,
-} from "./deps.test.ts";
+import { commands } from "../src/mod.ts";
+import { assert, assertRejects, describe, it } from "./deps.test.ts";
+import { getDummyCtx } from "./utils.ts";
 
 describe("commands", () => {
   it("should install the setMyCommands method on the context", () => {
-    const context = dummyCtx({});
+    const context = getDummyCtx({});
 
     const middleware = commands();
     middleware(context, async () => {});
@@ -27,7 +12,7 @@ describe("commands", () => {
     assert(context.setMyCommands);
   });
   it("should install the getNearestCommand method on the context", () => {
-    const context = dummyCtx({});
+    const context = getDummyCtx({});
 
     const middleware = commands();
     middleware(context, async () => {});
@@ -37,7 +22,7 @@ describe("commands", () => {
 
   describe("setMyCommands", () => {
     it("should throw an error if there is no chat", async () => {
-      const context = dummyCtx({ noMessage: true });
+      const context = getDummyCtx({ noMessage: true });
 
       const middleware = commands();
       middleware(context, async () => {});
@@ -50,28 +35,3 @@ describe("commands", () => {
     });
   });
 });
-
-export function dummyCtx({ userInput, language, noMessage }: {
-  userInput?: string;
-  language?: string;
-  noMessage?: boolean;
-}) {
-  const u = { id: 42, first_name: "yo", language_code: language } as User;
-  const c = { id: 100, type: "private" } as Chat;
-  const m = noMessage ? undefined : ({
-    text: userInput,
-    from: u,
-    chat: c,
-  } as Message);
-  const update = {
-    message: m,
-  } as Update;
-  const api = {
-    raw: { setMyCommands: spy(resolvesNext([true] as const)) },
-  } as unknown as Api;
-  const me = { id: 42, username: "bot" } as UserFromGetMe;
-  const ctx = new Context(update, api, me) as CommandsFlavor<Context>;
-  const middleware = commands();
-  middleware(ctx, async () => {});
-  return ctx;
-}
